requirements.txt
pandas
numpy
matplotlib
jupyter
python-dotenv
sqlalchemy 
pymysql